requirement.txt
discord.py
requests
python-dotenv
asyncio
openai
beautifulsoup4
lxml
PyPDF2
pycryptodome
tweepy
pandas
